refactor: Moving users to the new architecture
