+++ src/modules/User/routes/index.ts
@@ -5,17 +5,17 @@ export default [{
     path: "/users",
     name: "users",
     meta: { middleware: [auth, admin] },
-    component: () => import("@/modules/User/views/Index.vue").then(m => m.default)
+    component: () => import("@/modules/User/views/Datagrid/Index.vue").then(m => m.default)
 }, {
     path: "/users/create",
     name: "userCreate",
     meta: { middleware: [auth, admin] },
-    component: () => import("@/modules/User/views/CreateOrEdit.vue").then(m => m.default),
+    component: () => import("@/modules/User/views/CreateOrEdit/Index.vue").then(m => m.default),
     props: true
 }, {
     path: "/users/edit/:id(\\d+)",
     name: "userEdit",
     meta: { middleware: [auth, admin] },
-    component: () => import("@/modules/User/views/CreateOrEdit.vue").then(m => m.default),
+    component: () => import("@/modules/User/views/CreateOrEdit/Index.vue").then(m => m.default),
     props: true
 }]
