+++ src/modules/User/views/Datagrid/useIndex.ts
@@ -2,7 +2,7 @@ import { reactive, onMounted } from "vue"
 import { onBeforeRouteUpdate } from "vue-router"
 import useTableGrid from "@/core/composables/useTableGrid"
 import useHttp from "@/core/composables/useHttp"
-import UserService from "../services"
+import UserService from "@/modules/User/services"
 
 type Params =  string | string[][] | Record<string, string> | URLSearchParams | undefined
